multi-meta-rag/requirements.txt
groq==0.13.0
sentence-transformers==3.3.1
torch==2.6.0
transformers==4.47.0
qdrant-client==1.11.1
arxiv==2.1.3
pymupdf==1.25.5
pypdf==4.3.1
langchain-text-splitters==0.3.0
streamlit==1.38.0
python-dotenv==1.0.1
requests==2.32.3
faiss-cpu==1.13.2